more renaming and moving of autocompound and swap related things
