+++ smart-contracts/contracts/cl-vault/src/test_tube/mod.rs
@@ -1,9 +1,9 @@
 mod admin;
 mod any_deposit;
+mod autocompound;
 mod deposit_withdraw;
 mod helpers;
 mod initialize;
 mod proptest;
 mod range;
-mod redeposit;
 mod rewards;
