+++ smart-contracts/contracts/cl-vault/src/vault/autocompound.rs
@@ -1,17 +1,95 @@
+use cosmwasm_std::Order;
+use cosmwasm_std::{
+    to_json_binary, DepsMut, Env, MessageInfo, Response, SubMsg, SubMsgResult, Uint128,
+};
+use osmosis_std::types::cosmos::bank::v1beta1::{Input, MsgMultiSend, Output};
+use osmosis_std::types::osmosis::concentratedliquidity::v1beta1::ConcentratedliquidityQuerier;
+use osmosis_std::types::osmosis::concentratedliquidity::v1beta1::MsgCreatePositionResponse;
+
 use crate::error::ContractResult;
 use crate::helpers::{get_unused_balances, must_pay_one_or_two_from_balance};
 use crate::msg::{ExecuteMsg, MergePositionMsg};
 use crate::reply::Replies;
-use crate::state::{POOL_CONFIG, POSITION};
+use crate::rewards::CoinList;
+use crate::state::USER_REWARDS;
+use crate::state::{MigrationStatus, MIGRATION_STATUS, POOL_CONFIG, POSITION};
 use crate::vault::concentrated_liquidity::create_position;
 use crate::ContractError;
-use cosmwasm_std::{
-    to_json_binary, DepsMut, Env, MessageInfo, Response, SubMsg, SubMsgResult, Uint128,
-};
-use osmosis_std::types::osmosis::concentratedliquidity::v1beta1::ConcentratedliquidityQuerier;
-use osmosis_std::types::osmosis::concentratedliquidity::v1beta1::MsgCreatePositionResponse;
 
-/// Execute the redeposit process, creating a new position with unused balances.
+// Migration is a to-depreacate entrypoint useful to migrate from Distribute to Accumulate after Autocompound implementation
+pub fn execute_migration_step(
+    deps: DepsMut,
+    env: Env,
+    amount_of_users: Uint128,
+) -> Result<Response, ContractError> {
+    let mut migration_status = MIGRATION_STATUS.load(deps.storage)?;
+
+    if matches!(migration_status, MigrationStatus::Closed) {
+        return Err(ContractError::MigrationStatusClosed {});
+    }
+
+    let mut outputs = Vec::new();
+    let mut addresses = Vec::new();
+    let mut total_amount = CoinList::new();
+
+    // Iterate user rewards in a paginated fashion
+    for item in USER_REWARDS
+        .range(deps.storage, None, None, Order::Ascending)
+        .take(amount_of_users.u128() as usize)
+    {
+        let (address, rewards) = item?;
+        deps.api.debug(format!("address {:?}", address).as_str());
+        deps.api.debug(format!("rewards {:?}", rewards).as_str());
+
+        addresses.push(address.clone());
+        outputs.push(Output {
+            address: address.to_string(),
+            coins: rewards.osmo_coin_from_coin_list(),
+        });
+        total_amount.add(rewards)?;
+    }
+    deps.api
+        .debug(format!("total_amount {:?}", total_amount).as_str());
+
+    // Remove processed rewards in a separate iteration.
+    for addr in addresses {
+        USER_REWARDS.remove(deps.storage, addr);
+    }
+
+    // Check if this is the last execution.
+    let is_last_execution = USER_REWARDS
+        .range(deps.storage, None, None, Order::Ascending)
+        .next()
+        .is_none();
+    deps.api
+        .debug(format!("is_last_execution {:?}", is_last_execution).as_str());
+
+    if is_last_execution {
+        deps.api.debug("{:?}");
+        migration_status = MigrationStatus::Closed;
+        MIGRATION_STATUS.save(deps.storage, &migration_status)?;
+    }
+
+    let mut response = Response::new();
+    // Only if there are rewards append the send_message
+    if !total_amount.is_empty() {
+        let send_message = MsgMultiSend {
+            inputs: vec![Input {
+                address: env.contract.address.to_string(),
+                coins: total_amount.osmo_coin_from_coin_list(),
+            }],
+            outputs,
+        };
+        response = response.add_message(send_message);
+    }
+    response = response
+        .add_attribute("migration_status", format!("{:?}", migration_status))
+        .add_attribute("is_last_execution", is_last_execution.to_string());
+
+    Ok(response)
+}
+
+/// Execute the autocompound process, creating a new position with unused balances.
 ///
 /// # Arguments
 ///
@@ -25,7 +103,7 @@ use osmosis_std::types::osmosis::concentratedliquidity::v1beta1::MsgCreatePositi
 ///
 /// # Returns
 ///
-/// Returns a `Response` containing the result of the redeposit operation.
+/// Returns a `Response` containing the result of the autocompound operation.
 pub fn execute_autocompound(
     deps: DepsMut,
     env: Env,
@@ -69,21 +147,21 @@ pub fn execute_autocompound(
             Replies::Redeposit as u64,
         ))
         .add_attribute("method", "execute")
-        .add_attribute("action", "redeposit")
+        .add_attribute("action", "autocompound")
         .add_attribute("lower_tick", format!("{:?}", position.lower_tick))
         .add_attribute("upper_tick", format!("{:?}", position.upper_tick))
         .add_attribute("token0", format!("{:?}", token0.clone()))
         .add_attribute("token1", format!("{:?}", token1.clone())))
 }
 
-/// Handle the reply from the redeposit operation and then calling merge position
+/// Handle the reply from the autocompound operation and then calling merge position
 /// on the newly created position.
 ///
 /// # Arguments
 ///
 /// * `deps` - Dependencies for interacting with the contract.
 /// * `env` - Environment for fetching contract address.
-/// * `data` - Result of the redeposit operation.
+/// * `data` - Result of the autocompound operation.
 ///
 /// # Errors
 ///
@@ -92,7 +170,7 @@ pub fn execute_autocompound(
 /// # Returns
 ///
 /// Returns a `Response` containing the result of the merge operation.
-pub fn handle_redeposit_reply(
+pub fn handle_autocompound_reply(
     deps: DepsMut,
     env: Env,
     data: SubMsgResult,
@@ -117,7 +195,7 @@ pub fn handle_redeposit_reply(
             Replies::Merge.into(),
         ))
         .add_attribute("method", "reply")
-        .add_attribute("action", "handle_redeposit_reply")
+        .add_attribute("action", "handle_autocompound_reply")
         .add_attribute(
             "position_ids",
             format!("{:?}", vec![create_position_message.position_id]),
