+++ smart-contracts/contracts/cl-vault/src/vault/mod.rs
@@ -1,9 +1,9 @@
 pub mod admin;
 pub mod any_deposit;
+pub mod autocompound;
 pub mod concentrated_liquidity;
 pub mod exact_deposit;
 pub mod merge;
 pub mod range;
-pub mod redeposit;
 mod swap;
 pub mod withdraw;
